Refs #23. introduce AbstractImage and AbstractImageSeries. still ongoing
