+++ python/ivenus/AbstractImage.py
@@ -1,0 +1,10 @@
+# interface for Image data object
+
+class AbstractImage:
+
+    def getData(self):
+        raise NotImplementedError
+        
+        
+    def save(self):
+        raise NotImplementedError
