+++ python/ivenus/AbstractImageSeries.py
@@ -1,0 +1,13 @@
+# interface for image series
+
+class AbstractImageSeries:
+    
+    
+    def getImage(self, index, **kwds):
+        
+    
+        
+    def getData(self, index):
+        path = self.getFilename(index)
+        return ImageFile(path).getData()
+
